Plate number recognition/requirements.txt
numpy==1.18.5
tensorflow==2.3.1
tensorflow-estimator==2.3.0
opencv-contrib-python==4.5.1.48
jupyter
pandas
matplotlib==3.4.1
ipympl
pillow==8.2.0
